+++ .tape.js
@@ -1,0 +1,16 @@
+module.exports = {
+	'postcss-transform-shortcut': {
+		'rotate': {
+			message: 'supports rotate property'
+		},
+		'scale': {
+			message: 'supports scale property'
+		},
+		'translate': {
+			message: 'supports translate property'
+		},
+		'combo': {
+			message: 'supports multiple shortcut properties'
+		}
+	}
+};