+++ CHANGELOG.md
@@ -1,10 +1,17 @@
-## 1.0.0 (2015-07-18)
+# Changes to Transform Shortcut
+
+### 2.0.0 (November 27, 2016)
+
+- Updated: Correct `rotate` assignment
+- Updated: boilerplate conventions (Node v6.9.1 LTS)
+
+### 1.0.0 (July 18, 2015)
 
 Updated: PostCSS 5
 Updated: Develop dependencies
 Updated: ESLint configuration
 
-## 0.1.0 (2015-07-18)
+### 0.1.0 (July 18, 2015)
 
 Added: Initial release
 Added: Reference to specification
@@ -12,6 +19,6 @@ Added: Support for `.process` method
 Updated: Functionality to match specification
 Updated: Tests
 
-## 0.0.1 (2015-07-18)
+### 0.0.1 (July 18, 2015)
 
 Added: Pre-release